:sparkles: [BE] 아이템 사용, 먹방 횟수 추가 스케줄러 개발 완료
- 1시간마다 스케줄링 하는거로 변경도했는데 테스트 해봐야함

+++ scheduler-server/SchedulerServer/src/main/java/com/exodia/schedulerserver/db/repository/InGameClearLogRepository.java
@@ -1,5 +1,6 @@
 package com.exodia.schedulerserver.db.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -7,7 +8,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import com.exodia.schedulerserver.db.entity.InGameClearLog;
 import com.exodia.schedulerserver.db.entity.InGameCraftLog;
 
-public interface InGameClearLogRepository extends MongoRepository<InGameClearLog, Long> {
+public interface InGameClearLogRepository extends MongoRepository<InGameClearLog, BigInteger> {
 
 	InGameClearLog findByUserIdAndGameId(Long userId, String gameId);
 }
