+++ scheduler-server/SchedulerServer/src/main/java/com/exodia/schedulerserver/db/repository/InGameCraftLogRepository.java
@@ -1,13 +1,14 @@
 package com.exodia.schedulerserver.db.repository;
 
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.exodia.schedulerserver.db.entity.InGameCraftLog;
 
-public interface InGameCraftLogRepository extends MongoRepository<InGameCraftLog, Long> {
+public interface InGameCraftLogRepository extends MongoRepository<InGameCraftLog, BigInteger> {
 
 	int countByUserIdAndGameId(Long userId, String gameId);
