+++ scheduler-server/SchedulerServer/src/main/java/com/exodia/schedulerserver/db/repository/InGameHuntLogRepository.java
@@ -1,5 +1,6 @@
 package com.exodia.schedulerserver.db.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -7,6 +8,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import com.exodia.schedulerserver.db.entity.InGameCraftLog;
 import com.exodia.schedulerserver.db.entity.InGameHuntLog;
 
-public interface InGameHuntLogRepository extends MongoRepository<InGameHuntLog, Long> {
+public interface InGameHuntLogRepository extends MongoRepository<InGameHuntLog, BigInteger> {
 	List<InGameHuntLog> findByUserIdAndGameId(Long userId, String gameId);
 }
