+++ scheduler-server/SchedulerServer/src/main/java/com/exodia/schedulerserver/db/repository/InGameQuestLogRepository.java
@@ -3,6 +3,8 @@ package com.exodia.schedulerserver.db.repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import com.exodia.schedulerserver.db.entity.InGameQuestLog;
 
-public interface InGameQuestLogRepository extends MongoRepository<InGameQuestLog, Long> {
+import java.math.BigInteger;
+
+public interface InGameQuestLogRepository extends MongoRepository<InGameQuestLog, BigInteger> {
 	int countByUserIdAndGameId(Long userId, String gameId);
 }
